+++ java/org/rocksdb/RestoreBackupableDB.java
@@ -79,7 +79,7 @@ public class RestoreBackupableDB extends RocksObject {
    * @param backupId of backup to delete.
    * @throws org.rocksdb.RocksDBException
    */
-  public void deleteBackup(long backupId) throws RocksDBException {
+  public void deleteBackup(int backupId) throws RocksDBException {
     deleteBackup0(nativeHandle_, backupId);
   }
 
@@ -111,7 +111,7 @@ public class RestoreBackupableDB extends RocksObject {
       throws RocksDBException;
   private native void purgeOldBackups0(long nativeHandle, int numBackupsToKeep)
       throws RocksDBException;
-  private native void deleteBackup0(long nativeHandle, long backupId)
+  private native void deleteBackup0(long nativeHandle, int backupId)
       throws RocksDBException;
   protected native List<BackupInfo> getBackupInfo(long handle);
   private native void dispose(long nativeHandle);
